+++ packages/archives/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	* `@otuekong-portfolio/infrastructure-server`
 
 ### Changed
+* Updating `Alert` in `ContactFormCard` to auto-dismiss on succesful email sent
 * Applying 'Chip' to 'Pill' changes
 * Renaming `TechnologyChips` component to `TechnologyPills`
 * Updating `BioSection` component to display both bio text and `children`, when provided